+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Clixon Changelog
 
-* [7.4.0](#730) Planned: April 2025
+* [7.4.0](#740) Planned: April 2025
 * [7.3.0](#730) 30 January 2025
 * [7.2.0](#720) 28 October 2024
 * [7.1.0](#710) 3 July 2024
@@ -19,6 +19,8 @@ Planned: April 2025
 ### Features
 
 * Added new `ca_userdef` callback
+* New `clixon-restconf@2025-02-01.yang` revision
+  * Added timeout parameter
 
 ## 7.3.0
 30 January 2025